converted external numpy to pandas

+++ src/blop/digestion/__init__.py
@@ -1,5 +1,5 @@
 import pandas as pd
-
+import numpy as np
 from ..utils import get_beam_stats
 
 
@@ -9,7 +9,7 @@ def default_digestion_function(df: pd.DataFrame) -> pd.DataFrame:
 
 def beam_stats_digestion(df: pd.DataFrame, image_key, **kwargs) -> pd.DataFrame:
     for index, entry in df.iterrows():
-        stats = get_beam_stats(entry.loc[image_key], **kwargs)
+        stats = get_beam_stats(entry.loc[image_key].astype(dtype = np.float64), **kwargs)
     
         for attr, value in stats.items():
             if attr not in ["bbox"]:
